Consistently check if test kind is debug
